requirements.txt
# 핵심 Flask 라이브러리
Flask==3.1.1
Flask-SQLAlchemy
Flask-Migrate
Flask-Compress
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0

# 보안 및 인증
bcrypt==4.3.0
cryptography==45.0.6
cffi==1.17.1
pycparser==2.22

# 데이터베이스
oracledb==3.3.0

# 머신러닝 (실제 사용)
catboost==1.2.8
xgboost==3.0.2
pandas==2.3.1
numpy==2.1.3
joblib==1.5.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
six==1.17.0
threadpoolctl==3.6.0

# HTTP 요청 및 네트워크
requests==2.32.4
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.2
idna==3.10

# 설정 및 환경
python-dotenv==1.1.1
packaging==25.0
typing_extensions==4.14.1

# 컬러 출력 (개발/디버깅용)
colorama==0.4.6

# 벡터 데이터베이스  
chromadb>=0.4.0
PyPDF2==3.0.1

# 기타 의존성
cloudpickle==3.1.1
narwhals==2.0.1
numba==0.61.2
llvmlite==0.44.0
pillow==11.3.0
slicer==0.0.8
tqdm==4.67.1